% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_rsd_summaries.R
\name{calc_sd_rsd}
\alias{calc_sd_rsd}
\title{calculate values from summaries}
\usage{
calc_sd_rsd(data, low_cut, hi_cut = NULL)
}
\arguments{
\item{data}{data.frame of means and variances}

\item{low_cut}{means <= this value used for average sd}

\item{hi_cut}{means >= this value used for average rsd}
}
\value{
vector
}
\description{
given a data.frame of means and variances, calculate mean sd at low end and
mean rsd at high end.
}

