% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visqc.R
\name{calc_sd_rsd_nls}
\alias{calc_sd_rsd_nls}
\title{calculate values from summaries v2}
\usage{
calc_sd_rsd_nls(data, ...)
}
\arguments{
\item{data}{data.frame of means and variances}

\item{...}{other nls parameters}
}
\value{
vector
}
\description{
given a data.frame of means and variances, use a two step non-linear least squares.
The first step is done on the mean vs sd, then the estimates are used in a second
that estimates them using the mean vs rsd.
}

