% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{calculate_weights}
\alias{calculate_weights}
\title{calculate weights}
\usage{
calculate_weights(nonzero_loc, not_both = FALSE)
}
\arguments{
\item{nonzero_loc}{the non-zero location logical matrix}

\item{not_both}{consider things FALSE in both as information}
}
\value{
list
}
\description{
calculates the weights for correlations
}

