% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{correspondence}
\alias{correspondence}
\title{correspondence}
\usage{
correspondence(in_x, in_y)
}
\arguments{
\item{in_x}{which things are in X}

\item{in_y}{which things are in Y}
}
\value{
numeric
}
\description{
correspondence
}

