% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_features.R
\name{count_matching_features}
\alias{count_matching_features}
\title{matching features}
\usage{
count_matching_features(feature_matrix, zero_value = NA, groups = NULL)
}
\arguments{
\item{feature_matrix}{the feature to sample matrix.}

\item{zero_value}{what is the zero value? Default is NA}

\item{groups}{what are the groups}
}
\value{
data.frame
}
\description{
For a given set of feature-sample matrices, calculates how many features are
in that sample, as well as in common to all other samples, and if provided,
in common within and outside the same sample group.
}
