% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{determine_outliers}
\alias{determine_outliers}
\title{determine outliers}
\usage{
determine_outliers(
  median_correlations = NULL,
  outlier_fraction = NULL,
  cor_weight = 1,
  frac_weight = 1
)
}
\arguments{
\item{median_correlations}{median correlations}

\item{outlier_fraction}{outlier fractions}

\item{cor_weight}{how much weight for the correlation score?}

\item{frac_weight}{how much weight for the outlier fraction?}
}
\value{
data.frame
}
\description{
determine outliers
}
\details{
For outlier sample detection, one should 
  first generate median correlations using
  `median_correlations`, and outlier fractions using
  `outlier_fraction`. If you only have one or the other,
  than you should use named arguments to only pass the one
  or the other. 
  
  Alternatively, you can change the weighting used for median correlations
  or outlier fraction, including setting them to 0.
}
