% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_filtering.R
\name{filter_non_zero_percentage}
\alias{filter_non_zero_percentage}
\title{find non-zero by percentage}
\usage{
filter_non_zero_percentage(data_matrix, sample_classes, keep_perc = 0.75)
}
\arguments{
\item{data_matrix}{the matrix of values to work with}

\item{sample_classes}{the classes of each sample}

\item{keep_perc}{what percentage of samples in each class need a non-zero value}
}
\value{
matrix
}
\description{
Given a value matrix (features are rows, samples are columns), and sample classes,
find those things that are not zero in at least a certain
number of one of the classes
}

