% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/visqc.R
\name{gg_heatmap}
\alias{gg_heatmap}
\title{heatmap}
\usage{
gg_heatmap(dataMatrix, cols = NULL, limits = NULL)
}
\arguments{
\item{dataMatrix}{a data matrix with values for the heatmap}

\item{cols}{the colors to use}

\item{limits}{the limits to use for the data}
}
\value{
ggplot2 object
}
\description{
generates a heatmap using ggplot2
}

