% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{grp_cor}
\alias{grp_cor}
\title{10 sample to sample correlations}
\format{matrix with 10 rows and 10 columns, with row and colnames}
\source{
generated by rmflight
}
\description{
test data of 10 sample to sample correlations where samples are drawn
from two groups. Generated by rmflight from random distributions
}

