% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\docType{data}
\name{grp_cor_data}
\alias{grp_cor_data}
\title{grp_cor_data}
\format{List with 2 entries, data and class}
\source{
Robert M Flight
}
\usage{
grp_cor_data
}
\description{
Example data used for demonstrating \code{median correlation}. A list with
2 named entries:
}
\details{
\describe{
  \item{data}{a data matrix with 100 rows and 20 columns}
  \item{class}{a character vector of 20 entries denoting classes}
}

The data comes from two groups of samples, where there is ~0.85 correlation
within each group, and ~0.53 correlation between groups.
}
\keyword{datasets}

