% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\docType{data}
\name{grp_exp_data}
\alias{grp_exp_data}
\title{grp_exp_data}
\format{List with 2 entries, data and class}
\source{
Robert M Flight
}
\usage{
grp_exp_data
}
\description{
Example data that requires log-transformation before doing PCA or other QC.
A list with 2 named entries:
}
\details{
\describe{
  \item{data}{a data matrix with 1000 rows and 20 columns}
  \item{class}{a character vector of 20 entries denoting classes}
}

The data comes from two groups of samples, where there is ~0.80 correlation
within each group, and ~0.38 correlation between groups.
}
\keyword{datasets}

