% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{grp_info}
\alias{grp_info}
\title{10 sample meta-data}
\format{data.frame with 10 rows, and 2 columns, \code{grp} defining with group
and \code{set}, defining the set.}
\source{
generated by rmflight
}
\description{
meta-data for \code{grp_cor}.
}

