% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ici_kendallt}
\alias{ici_kendallt}
\title{Calculates ici-kendall-tau}
\usage{
ici_kendallt(x, y, perspective = "local", output = "simple")
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}

\item{perspective}{should we consider the "local" or "global" perspective?}
}
\value{
kendall tau correlation
}
\description{
Calculates ici-kendall-tau
}
\details{
Calculates the information-content-informed Kendall-tau correlation measure.
  This correlation is based on concordant and discordant ranked pairs, like Kendall-tau,
  but also includes missing values (as NA). Missing values are assumed to be *primarily* due
  to lack of detection due to instrumental sensitivity, and therefore encode *some* information.
  
  For more details see the ICI-Kendall-tau vignette:
  \code{vignette("ici-kendalltau", package = "visualizationQualityControl")}
}
\examples{
data("grp_cor_data")
exp_data = grp_cor_data$data
x = exp_data[, 1]
y = exp_data[, 2]
kendallt(x, y)
cor(x, y, method = "kendall") 

x = sort(rnorm(100))
y = x + 1
y2 = y
y2[1:10] = NA
kendallt(x, y)
kendallt(x, y2, "global")
kendallt(x, y2)

}
