% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{information_volume}
\alias{information_volume}
\title{information volume}
\usage{
information_volume(in_x, in_y)
}
\arguments{
\item{in_x}{which things are in X}

\item{in_y}{which things are in Y}
}
\value{
numeric
}
\description{
calculates the information volume weight
}
\details{
The information volume is based on how many things are actually
  not missing in both samples. For weighting of correlations, this value
  will be scaled by the maximum observed value across calculated correlations.
}

