% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_filtering.R
\name{keep_non_zero_percentage}
\alias{keep_non_zero_percentage}
\title{keep features with percentage of non-zeros}
\usage{
keep_non_zero_percentage(data_matrix, sample_classes = NULL,
  keep_num = 0.75)
}
\arguments{
\item{data_matrix}{the matrix of values to work with}

\item{sample_classes}{the classes of each sample}

\item{keep_num}{what number of samples in each class need a non-zero value (see Details)}
}
\value{
matrix
}
\description{
Given a value matrix (features are columns, samples are rows), and sample classes, 
find those things that are \emph{not zero} in at least a certain 
number of samples in one of the classes, and keep those features for further
processing.
}
\details{
The number of samples that must be non-zero can be expressed either as a whole
  number (that is greater than one), or as a fraction that will be be multiplied 
  by the number of samples in each class to get the lower limits for each of the classes.
}

