% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kendallt_matrix}
\alias{kendallt_matrix}
\alias{ici_kendall_matrix}
\title{Calculates ici-kendall-tau matrix}
\usage{
ici_kendall_matrix(x, perspective = "local")
}
\arguments{
\item{x}{data matrix}

\item{perspective}{should we consider the "local" or "global" perspective?}
}
\value{
kendall tau correlation
}
\description{
Calculates ici-kendall-tau matrix
}
