% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{locally_it_weighted_pairwise_correlation}
\alias{locally_it_weighted_pairwise_correlation}
\title{locally information theoretic weighted pairwise correlation}
\usage{
locally_it_weighted_pairwise_correlation(
  data_matrix,
  use = "pairwise.complete.obs",
  exclude_na = TRUE,
  exclude_inf = TRUE,
  exclude_0 = FALSE,
  zero_value = 0,
  method = "pearson"
)
}
\arguments{
\item{data_matrix}{the data}

\item{use}{what data to use, see \code{cor}}

\item{exclude_na}{should NA values be excluded (default TRUE)}

\item{exclude_inf}{should Inf values be excluded (default TRUE)}

\item{exclude_0}{should 0 values be excluded (default FALSE)}

\item{zero_value}{what value represents zero (default is 0)}

\item{method}{which method of correlation to use}
}
\value{
list
}
\description{
given a data matrix, calculate \emph{inter-row} correlation values that are weighted
by information content where only the pairwise information (jaccard index) is considered.
}
\details{
The function returns a named list with:
  \describe{
    \item{cor}{the weighted correlation matrix}
    \item{keep}{a logical matrix indicating which points passed filtering}
    \item{raw}{the raw correlations}
    \item{jaccard}{the jaccard information index used to weight the values}
    }
}
