% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{median_class_correlations}
\alias{median_class_correlations}
\title{calculate median class correlations}
\usage{
median_class_correlations(cor_matrix, sample_classes = NULL)
}
\arguments{
\item{cor_matrix}{the sample - sample correlations}

\item{sample_classes}{the sample classes as a character or factor}
}
\value{
matrix
}
\description{
Given a correlation matrix the sample class information,
calculates the median correlations of the samples within
the class and between classes.
}
