% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{outlier_fraction}
\alias{outlier_fraction}
\title{fraction of outliers}
\usage{
outlier_fraction(data, sample_classes = NULL, n_trim = 3, n_sd = 5,
  remove_0 = FALSE)
}
\arguments{
\item{data}{the data matrix (samples are rows, columns are features)}

\item{sample_classes}{the sample classes}

\item{n_trim}{how many features to trim at each end (default is 3)}

\item{n_sd}{how many SD before treated as outlier (default is 5)}

\item{remove_0}{should zeros be removed before calculating? (default is TRUE)}
}
\value{
data.frame
}
\description{
Calculates the fraction of entries in each sample that are more than \code{X}
standard deviations from the trimmed mean. See Details.
}
\details{
Based on the Gerlinski paper \href{https://dx.doi.org/10.1093/bioinformatics/btv425}{link}
for each feature (in a sample class), take the range across all the samples,
remove the \code{n_trim} lowest and highest values, and calculate the \code{mean}
and \code{sd}, and the actual upper and lower ranges of \code{n_sd} from the
\code{mean}. For each sample and feature, determine if \emph{within} or \emph{outside}
that limit. Fraction is reported as the number of features outside the range.
}

