% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{pairwise_correlation}
\alias{pairwise_correlation}
\title{pairwise correlation}
\usage{
pairwise_correlation(data_matrix, use = "pairwise", exclude_na = TRUE,
  exclude_inf = TRUE, exclude_0 = FALSE, method = "pearson")
}
\arguments{
\item{data_matrix}{the data}

\item{use}{what data to use, "pairwise" or "complete"}

\item{exclude_na}{should NA values be excluded (default TRUE)}

\item{exclude_inf}{should Inf values be excluded (default TRUE)}

\item{exclude_0}{should 0 values be excluded (default FALSE)}

\item{method}{which method of correlation to use}
}
\value{
list
}
\description{
given a data matrix, calculate inter-row correlation values.
}
\details{
The function returns a named list with:
  \describe{
    \item{cor}{the correlation matrix}
    \item{count}{how many points were used in the correlation}
    \item{keep}{a logical matrix indicating which points passed filtering}
    }
}

