% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{pairwise_correlation_both0}
\alias{pairwise_correlation_both0}
\title{pairwise correlation keep both zero}
\usage{
pairwise_correlation_both0(data_matrix, use = "pairwise", exclude_na = TRUE,
  exclude_inf = TRUE, exclude_0 = FALSE, method = "pearson")
}
\arguments{
\item{data_matrix}{the data}

\item{use}{what data to use, "pairwise" or "complete"}

\item{exclude_na}{should NA values be excluded (default TRUE)}

\item{exclude_inf}{should Inf values be excluded (default TRUE)}

\item{exclude_0}{should 0 values be excluded (default FALSE)}

\item{method}{which method of correlation to use}
}
\value{
matrix
}
\description{
given a data matrix, calculate inter-row correlation values, but if both
entries have zero, keep it
}

