% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visqc.R
\name{pairwise_nonzero}
\alias{pairwise_nonzero}
\title{pairwise non-zero}
\usage{
pairwise_nonzero(data_matrix, use = "pairwise", exclude_na = TRUE,
  exclude_inf = TRUE)
}
\arguments{
\item{data_matrix}{the data}

\item{use}{how to work with the data}

\item{exclude_na}{should NA values be excluded (default TRUE)}

\item{exclude_inf}{should Inf values be excluded (default TRUE)}
}
\value{
matrix
}
\description{
given a data matrix, how many entries are non-zero on a pairwise basis
}

