% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{similarity_reorderbyclass}
\alias{similarity_reorderbyclass}
\title{reorder by sample class}
\usage{
similarity_reorderbyclass(similarity_matrix, sample_classes = NULL,
  transform = "none")
}
\arguments{
\item{similarity_matrix}{matrix of similarities between objects}

\item{sample_classes}{list of indices of sample classes}

\item{transform}{a transformation to apply to the data}
}
\value{
a \link{dendrogram} object. To get the order use \code{order.dendogram}.
}
\description{
to avoid spurious visualization problems, it is useful in a heatmap visualization
to reorder the samples within each sample class. This function uses
hierarchical clustering and \link{dendsort} to sort entries in a heatmap.
}
\examples{
set.seed(1234)
mat <- matrix(rnorm(100, 2, sd = 0.5), 10, 10)
rownames(mat) <- colnames(mat) <- letters[1:10]
neworder <- similarity_reorderbyclass(mat)
mat[order(neworder), order(neworder)]

sample_class <- data.frame(grp = rep(c("grp1", "grp2"), each = 5))
sample_class <- split(rownames(mat), sample_class$grp)
neworder2 <- similarity_reorderbyclass(mat, sample_class)

mat[neworder2, neworder2]
}

