% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_groups.R
\name{split_groups}
\alias{split_groups}
\title{split groups}
\usage{
split_groups(in_matrix, groups = NULL)
}
\arguments{
\item{in_matrix}{the matrix we want to split up}

\item{groups}{a data.frame, character vector or factor}
}
\value{
list of groups
}
\description{
Given a matrix and a data.frame, character vector or data.frame of groups,
splits the indices / names of the matrix into groups appropriately. This function
assumes that the matrix and the groups are in the correct order!!
}
