% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/visqc.R
\name{summarize_data}
\alias{summarize_data}
\title{summarize data}
\usage{
summarize_data(in_data, sample_classes = NULL, avg_function = mean)
}
\arguments{
\item{in_data}{matrix or data.frame}

\item{sample_classes}{which samples are in which class}

\item{avg_function}{which function to use for summary}
}
\value{
data.frame
}
\description{
summarizes a matrix or data.frame, where rows are samples
and columns are features
}

