% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principal_component_testing.R
\name{visqc_cor_pca_scores}
\alias{visqc_cor_pca_scores}
\title{correlate scores and outcome}
\usage{
visqc_cor_pca_scores(pca_scores, sample_info)
}
\arguments{
\item{pca_scores}{the scores matrix to test}

\item{sample_info}{data.frame of sample attributes to test

Important: All of the attributes must be numeric, or character.
  If character, they will be transformed to a factor, and the numeric
  factor levels will be used instead.
  If missing values are present, that is OK, as long as they are
  missing-not-at-random (i.e. missing at the low end of the values).}
}
\value{
data.frame
}
\description{
Given a matrix of PCA scores, set of sample attributes to test,
goes through and performs an ICI-Kt of the scores versus the attribute.
}
