% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendalltau.R
\name{visqc_it_kendallt}
\alias{visqc_it_kendallt}
\title{information-theoretic kendall tau}
\usage{
visqc_it_kendallt(
  data_matrix,
  exclude_na = TRUE,
  exclude_inf = TRUE,
  exclude_0 = TRUE,
  zero_value = 0,
  perspective = "local"
)
}
\arguments{
\item{data_matrix}{samples are rows, features are columns}

\item{exclude_na}{should NA values be treated as NA?}

\item{exclude_inf}{should Inf values be treated as NA?}

\item{exclude_0}{should zero values be treated as NA?}

\item{zero_value}{what is the actual zero value?}

\item{perspective}{how to treat missing data in denominator and ties, see details}
}
\value{
numeric
}
\description{
Given a data-matrix, computes the information-theoretic Kendall-tau-b between
all samples.
}
