% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principal_components.R
\name{visqc_pca}
\alias{visqc_pca}
\title{sensible principal components}
\usage{
visqc_pca(pca_decomp, princomps = c(1, 2), groups = NULL, labels = NULL,
  dot_size = 4)
}
\arguments{
\item{pca_decomp}{the results of \code{prcomp}}

\item{princomps}{which principal components to plot}

\item{groups}{factor defining the classes of each object}

\item{labels}{labels of the points}

\item{dot_size}{size of the points (default is 1)}
}
\description{
given the results of \code{\link{prcomp}}, generate decent looking 
plot of selected components using \code{ggbiplot}.
}

