% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principal_components.R
\name{visqc_score_contributions}
\alias{visqc_score_contributions}
\title{calculate pca contributions}
\usage{
visqc_score_contributions(pca_scores)
}
\arguments{
\item{pca_scores}{matrix of scores, columns are each PC}
}
\value{
data.frame
}
\description{
Given a set of PCA scores, calculates their variance contributions, cumulative
contributions, and generates a percent label that can be used for labeling plots.
}
