% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principal_component_testing.R
\name{visqc_test_pca_loadings}
\alias{visqc_test_pca_loadings}
\title{test loadings}
\usage{
visqc_test_pca_loadings(
  loadings,
  test_columns,
  progress = FALSE,
  direction = FALSE
)
}
\arguments{
\item{loadings}{matrix of loadings from pca decomposition}

\item{test_columns}{names of the columns of the loadings to test}

\item{progress}{should progress be reported}

\item{direction}{should direction of loading be tested?}
}
\value{
named list
}
\description{
Given a matrix of loadings for principal components, and a set of components to test,
for each loading in each component, generates a null distribution from the other
loadings in all the other components, and reports a p-value for that loading.
}
