% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_rsd_summaries.R
\name{calculate_fratio}
\alias{calculate_fratio}
\title{calculate F-ratio}
\usage{
calculate_fratio(data, data_classes)
}
\arguments{
\item{data}{the data matrix, with samples (columns) and features (rows)}

\item{data_classes}{what are the classes of the samples (columns)}
}
\value{
vector
}
\description{
given a data matrix of samples (columns) and features (rows), and a vector of classes (character or factor),
calculate an F-ratio for each feature.
}
