% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{correspondence}
\alias{correspondence}
\title{correspondence}
\usage{
correspondence(in_x, in_y, not_both = FALSE)
}
\arguments{
\item{in_x}{which things are in X}

\item{in_y}{which things are in Y}

\item{not_both}{also include those things NOT in both}
}
\value{
numeric
}
\description{
correspondence
}
\details{
Correspondence is based on how many things agree compared to how
  many things do not agree. So by default it is the ratio of
  
    things observed in both / things observed in either
    
  This should handle well most types of data, including sparse data sets.
  However, there are cases where the fact that values are missing in both
  things being compared counts as information. This alternative method is
  triggered by setting \code{not_both = TRUE}, resulting in the calculation:
  
    (things observed in both + things not observed in both) / total # of things
    
  This value will be scaled to the largest observed correspondence for weighting
  as well.
}

