% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{determine_outliers}
\alias{determine_outliers}
\title{determine outliers}
\usage{
determine_outliers(
  median_correlations,
  outlier_fraction,
  cor_weight = 1,
  frac_weight = 1
)
}
\arguments{
\item{median_correlations}{median correlations}

\item{outlier_fraction}{outlier fractions}
}
\value{
data.frame
}
\description{
determine outliers
}
