% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_filtering.R
\name{filter_non_zero_percentage}
\alias{filter_non_zero_percentage}
\title{keep features with percentage of non-zeros}
\usage{
filter_non_zero_percentage(data_matrix, sample_classes = NULL,
  keep_num = 0.75)
}
\arguments{
\item{data_matrix}{the matrix of values to work with}

\item{sample_classes}{the classes of each sample}

\item{keep_num}{what number of samples in each class need a non-zero value (see Details)}
}
\value{
matrix
}
\description{
Given a value matrix (features are columns, samples are rows), and sample classes, 
find those things that are not zero in at least a certain 
number of one of the classes, and keep them
}
\details{
This function is being deprecated and all code should use the
  \code{keep_non_zero_percentage} function instead.
}
\seealso{
keep_non_zero_percentage
}

