% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{generate_group_colors}
\alias{generate_group_colors}
\title{create set of disjoint colors}
\usage{
generate_group_colors(n_group)
}
\arguments{
\item{n_group}{how many groups should there be colors for}
}
\description{
When multiple sample classes need to be visualized on a heatmap, it is useful
to be able to distinguish them by color. This function generates a set of colors
for sample classes
}

