% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{median_correlations}
\alias{median_correlations}
\title{calculate median correlations}
\usage{
median_correlations(cor_matrix, sample_classes = NULL, between_classes = FALSE)
}
\arguments{
\item{cor_matrix}{the sample - sample correlations}

\item{sample_classes}{the sample classes as a character or factor}

\item{between_classes}{should the between class correlations be evaluated?}
}
\value{
data.frame
}
\description{
Given a correlation matrix and optionally the sample class information,
calculates the median correlations of each sample to all other samples in
the same class. May be useful for determining outliers.
}
\details{
The data.frame has 6 columns:
\describe{
  \item{med_cor}{the median correlation with other samples}
  \item{sample_id}{the sample id, either the rowname or an index}
  \item{sample_class}{the class of the sample. If not provided, set to "C1"}
  \item{compare_class}{the class of the other sample}
  \item{plot_class}{\code{sample_class::compare_class} for easy grouping}
  \item{group}{whether the median is \emph{within} or \emph{between} groups of samples}
}
}
