% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{pairwise_correlation}
\alias{pairwise_correlation}
\title{pairwise correlation}
\usage{
pairwise_correlation(data_matrix, use = "pairwise.complete.obs",
  exclude_na = TRUE, exclude_inf = TRUE, exclude_0 = FALSE,
  zero_value = 0, method = "pearson", weight = TRUE)
}
\arguments{
\item{data_matrix}{the data}

\item{use}{what data to use, see \code{cor}}

\item{exclude_na}{should NA values be excluded (default TRUE)}

\item{exclude_inf}{should Inf values be excluded (default TRUE)}

\item{exclude_0}{should 0 values be excluded (default FALSE)}

\item{zero_value}{what value represents zero (default is 0)}

\item{method}{which method of correlation to use}

\item{weight}{should the correlations be weighted by information content and correspondence?}

\item{not_both}{should correspondence include things FALSE in both?}
}
\value{
list
}
\description{
given a data matrix, calculate \emph{inter-row} correlation values.
}
\details{
The function returns a named list with:
  \describe{
    \item{cor}{the correlation matrix}
    \item{keep}{a logical matrix indicating which points passed filtering}
    \item{raw}{the raw correlations}
    \item{info}{the weighting calculated from information content}
    \item{correspondence}{the weighting from correspondence}
    }
}

