% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{similarity_reorder}
\alias{similarity_reorder}
\title{cluster and reorder}
\usage{
similarity_reorder(
  similarity_matrix,
  matrix_indices = NULL,
  transform = "none",
  hclust_method = "complete",
  dendsort_type = "min"
)
}
\arguments{
\item{similarity_matrix}{matrix of similarities}

\item{matrix_indices}{indices to reorder}

\item{transform}{should a transformation be applied to the data first}

\item{hclust_method}{which method for clustering should be used?}

\item{dendsort_type}{how should the reordering be done? (default is "min")}
}
\value{
a \link{dendrogram} object. To get the order use \code{order.dendogram}.
}
\description{
given a matrix (maybe a distance matrix), cluster and then re-order using
dendsort.
}
