% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{similarity_reorderbyclass}
\alias{similarity_reorderbyclass}
\title{reorder by sample class}
\usage{
similarity_reorderbyclass(similarity_matrix, sample_classes,
  transform = "none")
}
\arguments{
\item{similarity_matrix}{matrix of similarities between objects}

\item{sample_classes}{list of indices of sample classes}

\item{transform}{a transformation to apply to the data}
}
\description{
to avoid spurious visualization problems, it is useful in a heatmap visualization
to reorder the samples within each sample class. This function uses
hierarchical clustering and \link{dendsort} to sort entries in a heatmap.
}

