% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_rsd_summaries.R
\name{summarize_data}
\alias{summarize_data}
\title{summarize data}
\usage{
summarize_data(
  in_data,
  sample_classes = NULL,
  avg_function = mean,
  log_transform = FALSE,
  remove_missing = NA
)
}
\arguments{
\item{in_data}{matrix or data.frame}

\item{sample_classes}{which samples are in which class}

\item{avg_function}{which function to use for summary}

\item{log_transform}{apply a log-transform to the mean}

\item{remove_missing}{remove missing values before summarizing}
}
\value{
data.frame
}
\description{
summarizes a matrix or data.frame, where columns are samples
and rows are features
}
