% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visqc.R
\name{summarize_data}
\alias{summarize_data}
\title{summarize data}
\usage{
summarize_data(in_data, sample_classes = NULL, avg_function = mean,
  log_transform = FALSE, remove_zeros = FALSE)
}
\arguments{
\item{in_data}{matrix or data.frame}

\item{sample_classes}{which samples are in which class}

\item{avg_function}{which function to use for summary}

\item{log_transform}{apply a log-transform to the mean}

\item{remove_zeros}{remove zeros before summarizing}
}
\value{
data.frame
}
\description{
summarizes a matrix or data.frame, where rows are samples
and columns are features
}

