% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principal_component_testing.R
\name{visqc_test_pca_scores}
\alias{visqc_test_pca_scores}
\title{test scores and outcome}
\usage{
visqc_test_pca_scores(pca_scores, sample_info)
}
\arguments{
\item{pca_scores}{matrix of scores from a PCA decomposition}

\item{sample_info}{data.frame of sample attributes to test}
}
\value{
data.frame
}
\description{
Given a matrix of PCA scores, set of sample attributes to test,
goes through and performs an ANOVA of the scores versus the attribute.
}
